completely remove tests.utils
